requirements.txt
PySide6>=6.8.0
scikit-learn>=1.6.0
joblib>=1.4.0
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
pyinstaller>=6.12.0

